MAINT: refactor of admin commands
